Implemented the delete function using `windows-rs`
